adding a clone-all option

+++ dotfiles/cli.sh
@@ -1,7 +1,17 @@
 #!/usr/bin/env bash
 function bitcar_cli() {
     local workspace_dir="${BITCAR_WORKSPACE_DIR:-"$HOME/bitcar-repos"}"
-    [ -z "$1" ] && cd "$workspace_dir" && return
+
+    if [ -z "$1" ]; then
+       cd "$workspace_dir";
+       return
+    fi
+
+    if [ "$1" == "--clone-all" ] || [ "$2" == "--clone-all" ]; then
+       bitcar "$@" && cd "$workspace_dir"
+       return
+    fi
+
     local target="$HOME/.bitcar/.bitcar_target"
     bitcar "$@" && if [ -f "$target" ]; then cd "$(cat "$target")" && rm "$target"; fi
 }
